ui: study session visual polish — speaker icon, prompt label, first-letter hint
- Move SpeakerVector to shared ui/AppIcons.kt so study and word-details screens use the same stroke-style icon instead of the filled Material VolumeUp
- Prompt labels ("TRANSLATE TO DUTCH", "FILL IN THE BLANK") now match design: 11sp, Bold, 1.4sp letter-spacing, uppercase
- First-letter hint restyled as pill (CircleShape, surfaceContainerHigh) with split spans: first letter in onSurface/SemiBold, blanks in Monospace with 4sp letter-spacing
